Split Hand into private classes
