--- BasicSample/Assets/HandTracking/Scripts/Hand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfb2bfae7cada4641912d7acc25cf395
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
